BudgetAssistant-backend/pyproject.toml
[project]
name = "budgetassistant-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arrow==1.3.0",
    "django==5.1.2",
    "django-choice-enumfields==1.1.3",
    "django-cors-headers==4.6.0",
    "django-environ==0.11.2",
    "django-pandas==0.6.7",
    "django-silk>=5.3.2",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "drf-spectacular==0.28.0",
    "importlib-resources==6.4.5",
    "jsonpickle==4.0.1",
    "lxml==5.3.0",
    "model-bakery==1.20.0",
    "psycopg2-binary==2.9.9",
    "networkx==3.4.1",
    "numpy==2.1.2",
    "openapi-generator-cli[jdk4py]>=7.13.0.post0",
    "pandas==2.2.3",
    "parameterized==0.9.0",
    "polyfactory==2.17.0",
    "pydantic==2.9.2",
    "pytest-django==4.9.0",
    "whitenoise==6.9.0",
    "python-dotenv>=1.1.1",
]
